Allow configuration of mappings and fix tab logic
Changed `luagit.config` to return final configuration.
Set up binds based on user configuration.
Renamed `kind` to `method` in `open()`.
Made `open()` close lazygit or return to it when ran in lazygit and files edited by lazygit, respectively (removes the need for keybinds in the bridge).
Fixed mistake in tab logic that made it so the wrong method variable was set and neovim would close.
Changed check for being in the only tab to support moving the lazygit tab to 1 without closing others.